Ch13 ex.4: Insertion sort
Is in Pile.java because it was a lot more easy to  do it with ArrayList.
Ps: I don't know exactly how Github works yet, so maybe now there are two folders named "Cartas". Sorry in advance to whoever is reading this, especially if it's me a few years in the future.